Update full_description.txt
You use Markdown in places for formatting. It's well supported by IzzyOnDroid, but not supported at all by F-Droid. To ensure your description looks good on websites and in clients on both platforms, I recommend limiting it to plain text and simple HTML with tags that F-Droid has confirmed support for.

I also added punctuation marks at the end of lines on my own initiative, as the Russian language is, on average, longer than English, and without indenting the first line of paragraphs or using indents between paragraphs, sentences will run together on smartphone screens. They'll be difficult to read. This isn't a big deal for short lines, but here the principle of style consistency within a single text comes into play.